+++ poseidon-server/src/main/java/net/minecraft/server/BlockCrops.java
@@ -95,7 +95,7 @@ public class BlockCrops extends BlockFlower {
                     float f2 = world.random.nextFloat() * f1 + (1.0F - f1) * 0.5F;
                     float f3 = world.random.nextFloat() * f1 + (1.0F - f1) * 0.5F;
                     float f4 = world.random.nextFloat() * f1 + (1.0F - f1) * 0.5F;
-                    EntityItem entityitem = new EntityItem(world, (double) ((float) i + f2), (double) ((float) j + f3), (double) ((float) k + f4), new ItemStack(Item.SEEDS));
+                    EntityItem entityitem = new EntityItem(world, (float) i + f2, (float) j + f3, (float) k + f4, new ItemStack(Item.SEEDS));
 
                     entityitem.pickupDelay = 10;
                     world.addEntity(entityitem);
