+++ poseidon-server/src/main/java/net/minecraft/server/BlockSnow.java
@@ -15,7 +15,7 @@ public class BlockSnow extends Block {
     public AxisAlignedBB e(World world, int i, int j, int k) {
         int l = world.getData(i, j, k) & 7;
 
-        return l >= 3 ? AxisAlignedBB.b((double) i + this.minX, (double) j + this.minY, (double) k + this.minZ, (double) i + this.maxX, (double) ((float) j + 0.5F), (double) k + this.maxZ) : null;
+        return l >= 3 ? AxisAlignedBB.b((double) i + this.minX, (double) j + this.minY, (double) k + this.minZ, (double) i + this.maxX, (float) j + 0.5F, (double) k + this.maxZ) : null;
     }
 
     public boolean a() {
